Move ThreadPoolExecutor anon to named class
